started fixing -Wsign-compare warning messages
Import internal commit 3d766220a2cfd0109fb7e955576000aae946b2f1

+++ groups/bsl/bslstl/bslstl_bitset.t.cpp
@@ -180,7 +180,7 @@ bool verifyBitset(const bsl::bitset<N> obj, const char *expected)
 template <size_t N>
 bool verifyBitset(const bsl::bitset<N> obj, unsigned long expected, int verbose)
 {
-    for (int bitIndex = 0; bitIndex < N; ++bitIndex) {
+    for (size_t bitIndex = 0; bitIndex < N; ++bitIndex) {
         int expectedBit = 0;
 
         if (bitIndex < sizeof(unsigned long) * CHAR_BIT) {
